change scope of eth dispatch param
